删除课程和删除课程班接口完善

+++ src/main/java/cn/ken/student/rubcourse/controller/sys/CourseController.java
@@ -68,7 +68,7 @@ public class CourseController {
     
     @DeleteMapping("removeCourseClass")
     @ApiOperation("删除课程班")
-    public Result removeCourseClass(HttpServletRequest httpServletRequest, Integer id) {
+    public Result removeCourseClass(HttpServletRequest httpServletRequest, Long id) {
         return courseService.removeCourseClass(httpServletRequest, id);
     }
