+++ src/main/java/cn/ken/student/rubcourse/service/ICourseService.java
@@ -30,5 +30,5 @@ public interface ICourseService extends IService<Course> {
     
     Result removeCourse(HttpServletRequest httpServletRequest, String id);
 
-    Result removeCourseClass(HttpServletRequest httpServletRequest, Integer id);
+    Result removeCourseClass(HttpServletRequest httpServletRequest, Long id);
 }
